+++ src/plone/restapi/upgrades/to0004.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+
+def assign_get_users_permission(setup_context):
+    """Assign the 'plone.restapi: Access Plone user information' permission
+       to Managers by default.
+    """
+    setup_context.runImportStepFromProfile(
+        'profile-plone.restapi.upgrades:0004',
+        'rolemap',
+        run_dependencies=False,
+        purge_old=False)